src/third_party/WebKit/LayoutTests/fast/scrolling/scrollbar-mousedown-mouseup.html

<!DOCTYPE html>
<style>
#scrollme {
  width: 100px;
  height: 100px;
  overflow: auto;
}
#scrollme p {
  height: 1000px;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/gesture-util.js"></script>

<div id="scrollme"><p>This is a scrollable div.</p></div>
<script>
function initEventHandlers(element) {
  element.addEventListener('mousedown', handleEvent);
  element.addEventListener('mouseup', handleEvent);
  element.addEventListener('click', handleEvent);
}

window.events = [];
function handleEvent(e) {
  window.events.push(e);
}

promise_test(async () => {
  var d = document.querySelector('#scrollme');
  initEventHandlers(d);

  var x = d.offsetLeft + d.offsetWidth - 4;
  var y = d.offsetTop + 4;
  await mouseClickOn(x, y);
  await mouseClickOn(x, y, 'middle');
  await waitFor(() => {return window.events.length == 4;} );
  assert_equals(events[0].type, "mousedown");
  assert_equals(events[0].target.id, "scrollme");
  assert_equals(events[0].which, 1);
  assert_equals(events[1].type, "mouseup");
  assert_equals(events[1].target.id, "scrollme");
  assert_equals(events[1].which, 1);
  assert_equals(events[2].type, "mousedown");
  assert_equals(events[2].target.id, "scrollme");
  assert_equals(events[2].which, 2);
  assert_equals(events[3].type, "mouseup");
  assert_equals(events[3].target.id, "scrollme");
  assert_equals(events[3].which, 2);
});
</script>    
